fix(frontend): topic fetch for server
