+++ packages/frontend/src/fetchers/topic.ts
@@ -1,121 +1,5 @@
 import { sortByCountDesc } from '@/fetchers/utils'
 
-/* fetchTopic
- *   fetch topics with give slug in given terms & session
- */
-export type SpeechData = {
-  slug: string
-  summary: string
-  title: string
-  date: Date
-  legislativeYuanMember: {
-    legislator: {
-      name: string
-      slug: string
-      imageLink?: string
-      image?: {
-        imageFile: {
-          url: string
-        }
-      }
-    }
-  }
-}
-
-export type TopicData = {
-  slug: string
-  title: string
-  speechesCount?: number
-  speeches?: SpeechData[]
-  relatedTopics?: {
-    slug: string
-    title: string
-  }[]
-}
-
-export const fetchTopic = async ({
-  slug,
-  legislativeMeeting,
-  legislativeMeetingSession,
-}: {
-  slug: string
-  legislativeMeeting: number
-  legislativeMeetingSession: number[]
-}): Promise<TopicData> => {
-  const url = process.env.NEXT_PUBLIC_API_URL as string
-  const where = {
-    where: {
-      slug: slug,
-    },
-    speechesWhere: {
-      legislativeMeeting: {
-        term: {
-          equals: legislativeMeeting,
-        },
-      },
-      legislativeMeetingSession: {
-        term: {
-          in: legislativeMeetingSession,
-        },
-      },
-    },
-  }
-  const res = await fetch(url, {
-    signal: AbortSignal.timeout(30000),
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      query: `
-        query Topic($where: TopicWhereUniqueInput!, $speechesWhere: SpeechWhereInput!) {
-          topic(where: $where) {
-            speeches(where: $speechesWhere) {
-              slug
-              summary
-              title
-              date
-              legislativeYuanMember {
-                legislator {
-                  name
-                  slug
-                  imageLink
-                  image {
-                    imageFile {
-                      url
-                    }
-                  }
-                }
-              }
-              legislativeMeeting {
-                term
-              }
-              legislativeMeetingSession {
-                term
-              }
-            }
-            slug
-            title
-            speechesCount(where: $speechesWhere)
-            relatedTopics {
-              slug
-              title
-            }
-          }
-        }
-      `,
-      variables: where,
-    }),
-  })
-
-  if (!res.ok) {
-    throw new Error(`Failed to fetch topic data, slug: ${slug}`)
-  }
-
-  const data = await res.json()
-  return data?.data?.topic
-}
-
 /* fetchTopTopicsForLegislator
  *   fetch top 5 topic which given legislator has more speeches in given terms & session
  */
